Keywords
I created a table for keywords / business types. The table can be
populated and associated with businesses even after the site is online.
However, unlike asking a user how much they "like" something, business
types are going to be a binary relationship. The business is classified
by it or not.