Finally. Rating built into the display for businesses
There's code to print out the rating selection in viewlib (called
dumpRating).

The code needs refactoring, but getting a version off the ground is
critical at this point as far as I'm concerned.

Going forward, we should try and keep the directories decent. The
search director should contain everything needed to actually find the
relevant results to what the user wants. The view director should be
dedicated to actually getting stuff on the page. The handler directory
is a catch-all for callbacks from the ajax scripts. They basically
route control around. naturally, db contains everything that modifies /
updates the DB.