Built a rating engine that populates the database with the recommendation for each user for each business
I've implemented it as a celery daemon that runs periodically (using
celery's version of a cron job called celerybeat). Essentially, every
10 seconds, the task pulls ALL of the users' ratings from the database,
calculates a correlation between all of them, and then assigns a
recommendation for each user for each business based on this
correlation. It's not easy (or even possible) to communicate through
python's data structures when using this cron-job-like task, so I've
created a table in the database called Recommendation. The task *can*
communicate with the database.

This is actually good for future scaling. The "website" portion of the
code can run on any run-of-the-mill server. The asynchronous task of
figuring out correlation and recommendation is split from the website's
functionality and can even be migrated to Amazon's web services.